Move out TextRange and TraversalControl out of Node.ts
