+++ src/compiler/ast/common/TextRange.ts
@@ -1,0 +1,4 @@
+export interface TextRange {
+    getPos(): number;
+    getEnd(): number;
+}
