+++ src/compiler/ast/common/index.ts
@@ -5,3 +5,5 @@ export * from "./Node";
 export * from "./QualifiedName";
 export * from "./Scope";
 export * from "./SyntaxList";
+export * from "./TextRange";
+export * from "./TraversalControl";
